Moved ArrayOps to Workarounds
