+++ logdoctor/workarounds/CMakeLists.txt
@@ -1,4 +1,7 @@
 
 target_sources(LogDoctor PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/lib.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/ranges_enumerate.h
     ${CMAKE_CURRENT_SOURCE_DIR}/ranges_join.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/ranges_zip.h
 )
